--- src/main/java/subway/common/CommonService.java
@@ -1,24 +1,0 @@
-package subway.common;
-
-import subway.main.validation.CheckValidOption;
-import subway.view.InputView;
-import subway.view.MainOutputView;
-
-import java.util.List;
-
-public class CommonService {
-    public static char selectOption(List<Character> optionList, InputView inputView) {
-        char option;
-        while (true) {
-            try {
-                MainOutputView.askOptionChoice();
-                option = inputView.selectOption();
-                CheckValidOption.validation(option, optionList);
-                break;
-            } catch (IllegalArgumentException e) {
-                System.out.println(e.getMessage());
-            }
-        }
-        return option;
-    }
-}
