+++ src/main/java/subway/station/StationService.java
@@ -2,7 +2,6 @@ package subway.station;
 
 import subway.line.validation.CheckStationRegisteredLine;
 import subway.station.validation.CheckRegisteredStation;
-import subway.view.station.StationManagementView;
 
 import java.util.List;
 
@@ -11,27 +10,15 @@ public class StationService {
     private static final String NOT_EXIST = ERROR_PREFIX + "등록되지 않은 역입니다.";
 
     public static boolean addStation(String stationName) {
-        boolean isAdd = false;
-        try {
-            Station station = new Station(stationName);
-            StationRepository.addStation(station);
-            isAdd = true;
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-        }
-        return isAdd;
+        Station station = new Station(stationName);
+        StationRepository.addStation(station);
+        return true;
     }
 
     public static boolean deleteStation(String stationName) {
-        boolean isDelete = false;
-        try {
-            CheckRegisteredStation.validation(stationName);
-            CheckStationRegisteredLine.validation(stationName);
-            isDelete = StationRepository.deleteStation(stationName);
-        } catch (IllegalArgumentException e) {
-            System.out.println(e.getMessage());
-        }
-        return isDelete;
+        CheckRegisteredStation.validation(stationName);
+        CheckStationRegisteredLine.validation(stationName);
+        return StationRepository.deleteStation(stationName);
     }
 
     public static Station findStation(String stationName) {
@@ -42,9 +29,7 @@ public class StationService {
         return station;
     }
 
-    public static boolean printAllStation() {
-        List<Station> stations = StationRepository.stations();
-        StationManagementView.printAllStation(stations);
-        return true;
+    public static List<Station> AllStation() {
+        return StationRepository.stations();
     }
 }
